+++ test/cube-to-cube.js
@@ -339,13 +339,13 @@ describe('Operation between cubes', function () {
                 new GenericDimension('period', 'season', ['summer', 'winter']),
                 new GenericDimension('location', 'city', ['paris', 'toledo', 'tokyo']),
             ]);
-            cube.createStoredMeasure('antennas', {}, 'float32', 0);
+            cube.createStoredMeasure('antennas', {}, 'uint32', 0);
 
             const cube2 = new Cube([
                 new GenericDimension('period', 'season', ['winter']),
                 new GenericDimension('location', 'city', ['paris', 'tokyo']),
             ]);
-            cube2.createStoredMeasure('otherMeasure', {}, 'float32', 666);
+            cube2.createStoredMeasure('otherMeasure', {}, 'uint32', 666);
             cube.hydrateFromCube(cube2);
 
             assert.deepEqual(cube.getNestedObject('antennas'), {
@@ -359,14 +359,14 @@ describe('Operation between cubes', function () {
                 new GenericDimension('period', 'season', ['summer', 'winter']),
                 new GenericDimension('location', 'city', ['paris', 'toledo', 'tokyo']),
             ]);
-            cube.createStoredMeasure('antennas', {}, 'float32', 0);
+            cube.createStoredMeasure('antennas', {}, 'uint32', 0);
 
             const cube2 = new Cube([
                 new GenericDimension('period', 'season', ['winter']),
                 new GenericDimension('location', 'city', ['paris', 'tokyo']),
             ]);
-            cube2.createStoredMeasure('antennas', {}, 'float32', 666);
-            cube2.createStoredMeasure('otherMeasure', {}, 'float32', 888);
+            cube2.createStoredMeasure('antennas', {}, 'uint32', 666);
+            cube2.createStoredMeasure('otherMeasure', {}, 'uint32', 888);
 
             cube.hydrateFromCube(cube2);
 
@@ -381,13 +381,13 @@ describe('Operation between cubes', function () {
                 new GenericDimension('period', 'season', ['summer', 'winter']),
                 new GenericDimension('location', 'city', ['paris', 'toledo', 'tokyo']),
             ]);
-            cube.createStoredMeasure('antennas', {}, 'float32', 0);
+            cube.createStoredMeasure('antennas', {}, 'uint32', 0);
 
             const cube2 = new Cube([
                 new GenericDimension('period', 'season', ['winter']),
                 new GenericDimension('location', 'city', ['paris', 'tokyo']),
             ]);
-            cube2.createStoredMeasure('antennas', {}, 'float32', 0);
+            cube2.createStoredMeasure('antennas', {}, 'uint32', 0);
             cube2.setNestedObject('antennas', { winter: { paris: 32, tokyo: 53 } });
 
             cube.hydrateFromCube(cube2);
@@ -403,14 +403,14 @@ describe('Operation between cubes', function () {
                 new GenericDimension('period', 'season', ['summer', 'winter']),
                 new GenericDimension('location', 'city', ['paris', 'toledo', 'tokyo']),
             ]);
-            cube.createStoredMeasure('antennas', {}, 'float32', 0);
+            cube.createStoredMeasure('antennas', {}, 'uint32', 0);
 
             const cube2 = new Cube([
                 new GenericDimension('period', 'season', ['winter']),
                 new GenericDimension('location', 'city', ['tokyo', 'losangeles', 'paris']),
             ]);
 
-            cube2.createStoredMeasure('antennas', {}, 'float32', 0);
+            cube2.createStoredMeasure('antennas', {}, 'uint32', 0);
             cube2.setNestedObject('antennas', { winter: { tokyo: 1, losangeles: 2, paris: 3 } });
 
             cube.hydrateFromCube(cube2);
@@ -426,7 +426,7 @@ describe('Operation between cubes', function () {
                 new GenericDimension('period', 'season', ['summer', 'winter']),
                 new GenericDimension('location', 'city', ['paris', 'toledo', 'tokyo']),
             ]);
-            cube.createStoredMeasure('antennas', {}, 'float32', 0);
+            cube.createStoredMeasure('antennas', {}, 'uint32', 0);
 
             const cube2 = new Cube([
                 new GenericDimension('period', 'season', ['winter']),
@@ -434,7 +434,7 @@ describe('Operation between cubes', function () {
                 new GenericDimension('location', 'city', ['paris', 'tokyo']),
             ]);
 
-            cube2.createStoredMeasure('antennas', {}, 'float32', 0);
+            cube2.createStoredMeasure('antennas', {}, 'uint32', 0);
             cube2.setNestedObject('antennas', {
                 winter: {
                     a: { paris: 1, tokyo: 2 },
@@ -456,11 +456,11 @@ describe('Operation between cubes', function () {
                 new GenericDimension('period', 'season', ['summer', 'winter']),
                 new GenericDimension('location', 'city', ['paris', 'toledo', 'tokyo']),
             ]);
-            cube.createStoredMeasure('antennas', {}, 'float32', 0);
+            cube.createStoredMeasure('antennas', {}, 'uint32', 0);
 
             const cube2 = new Cube([new GenericDimension('period', 'season', ['winter'])]);
 
-            cube2.createStoredMeasure('antennas', {}, 'float32', 0);
+            cube2.createStoredMeasure('antennas', {}, 'uint32', 0);
             cube2.setNestedObject('antennas', { winter: 32 });
 
             cube.hydrateFromCube(cube2);
@@ -477,13 +477,13 @@ describe('Operation between cubes', function () {
                 new GenericDimension('location', 'city', ['paris', 'toledo', 'tokyo']),
             ]);
 
-            cube.createStoredMeasure('antennas', {}, 'float32', 0);
+            cube.createStoredMeasure('antennas', {}, 'uint32', 0);
 
             const cube2 = new Cube([
                 new TimeDimension('time', 'month', '2010-04', '2010-06'),
                 new GenericDimension('location', 'city', ['toledo']),
             ]);
-            cube2.createStoredMeasure('antennas', {}, 'float32', 0);
+            cube2.createStoredMeasure('antennas', {}, 'uint32', 0);
             cube2.setNestedObject('antennas', {
                 '2010-04': { toledo: 1 },
                 '2010-05': { toledo: 2 },
@@ -504,13 +504,13 @@ describe('Operation between cubes', function () {
                 new TimeDimension('time', 'month', '2010-01', '2010-06'),
                 new GenericDimension('location', 'city', ['paris', 'toledo', 'tokyo']),
             ]);
-            cube.createStoredMeasure('antennas', {}, 'float32', 0);
+            cube.createStoredMeasure('antennas', {}, 'uint32', 0);
 
             const cube2 = new Cube([
                 new TimeDimension('time', 'quarter', '2010-Q2', '2010-Q2'),
                 new GenericDimension('location', 'city', ['toledo']),
             ]);
-            cube2.createStoredMeasure('antennas', {}, 'float32', 0);
+            cube2.createStoredMeasure('antennas', {}, 'uint32', 0);
             cube2.setNestedObject('antennas', { '2010-Q2': { toledo: 100 } });
 
             cube.hydrateFromCube(cube2);
